Generate Excel File
Added Pandas library to create DataFrames from the raw_data and dark_times dicts. Then appended that data to an excel file. Would like to format the data better so it outputs to excel in a better way. I tried a few things that almost worked, but didn't so there's probably a better way to do things. See comments on lines 273-280.