Fixed logic problem
Fixed logic problem where on May 21, the DarkTimes output is setting the start time to 10:41 PM (Dusk), when it should be setting the start time to 11:22 PM (Moon Set) by adding if not statement in the
else: # moon looks like |___---___|

Tested with Atascadero LAT/LON and appears to work, but displays in MDT time. Need to fix display tz.